Changed RMSD to double precision

+++ metrics/RMSD.h
@@ -30,25 +30,25 @@ class RMSD: public Metric{
 
 typedef struct
 {
-  float m[4][4];
+  double m[4][4];
 } MATRIX;
 
 #define vdiff2(a,b) ( ((a)[0]-(b)[0]) * ((a)[0]-(b)[0]) +	\
 		((a)[1]-(b)[1]) * ((a)[1]-(b)[1]) + \
 		((a)[2]-(b)[2]) * ((a)[2]-(b)[2]) )
 
-double alignedrmsd(float *v1, float *v2, int N);
-void centroid(float *ret, float *v, int N);
-int getalignmtx(float *v1, float *v2, int N, MATRIX *mtx);
-void crossproduct(float *ans, float *pt1, float *pt2);
+double alignedrmsd(double *v1, double *v2, int N);
+void centroid(double *ret, double *v, int N);
+int getalignmtx(double *v1, double *v2, int N, MATRIX *mtx);
+void crossproduct(double *ans, double *pt1, double *pt2);
 void mtx_root(MATRIX *mtx);
 int almostequal(MATRIX *a, MATRIX *b);
-void mulpt(MATRIX *mtx, float *pt);
+void mulpt(MATRIX *mtx, double *pt);
 void mtx_mul(MATRIX *ans, MATRIX *x, MATRIX *y);
 void mtx_identity(MATRIX *mtx);
-void mtx_trans(MATRIX *mtx, float x, float y, float z);
-int mtx_invert(float *mtx, int N);
-float absmaxv(float *v, int N);
+void mtx_trans(MATRIX *mtx, double x, double y, double z);
+int mtx_invert(double *mtx, int N);
+double absmaxv(double *v, int N);
 
 /*
    calculate rmsd between two structures
@@ -60,7 +60,7 @@ Returns: rmsd score
 Notes: mtx can be null. Transform will be rigid. Inputs must
 be previously aligned for sequence alignment
 */
-double rmsd(float *v1, float *v2, int N, float *mtx);
+double rmsd(double *v1, double *v2, int N, double *mtx);
 
 
 }
